# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Valheim Server UI API
  version: 1.0.0
  description: |
    HTTP API of valheim-server-ui. All endpoints are under /api/v1.
    Roles: viewer < operator < admin. Each operation declares the minimum role in `x-role`.
    Mutating requests must send header `X-Requested-With: valheim-ui` (CSRF guard).
    Long-running operations return 202 with a Job; follow progress via /events (SSE).
servers:
  - url: /api/v1
...
paths:
...
  /auth/setup:
    post:
      tags: [auth]
      summary: Create the first admin (only while no users exist)
      x-role: public
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetupRequest' }
      responses:
        '201': { $ref: '#/components/responses/UserResponse' }
        '404': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /auth/login:
    post:
      tags: [auth]
      summary: Local login; sets the session cookie
      x-role: public
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { $ref: '#/components/responses/UserResponse' }
        '401': { $ref: '#/components/responses/Error' }
        '423': { $ref: '#/components/responses/Error' }
...
  /auth/me:
    get:
      tags: [auth]
      x-role: viewer
      responses:
        '200': { $ref: '#/components/responses/UserResponse' }
        '401': { $ref: '#/components/responses/Error' }
  /auth/password:
    put:
      tags: [auth]
      summary: Change own password (local accounts only)
      x-role: viewer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePasswordRequest' }
      responses:
        '204': { description: Changed }
        '401': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /auth/oidc/login:
    get:
      tags: [auth]
      summary: Start OIDC authorization code flow (302 to provider)
      x-role: public
      parameters:
        - in: query
          name: next
          schema: { type: string }
          description: SPA path to return to after login (must start with /)
      responses:
        '302': { description: Redirect to provider }
        '404': { $ref: '#/components/responses/Error' }
...
  /instances/{instanceId}/backups:
    parameters: [{ $ref: '#/components/parameters/instanceId' }]
    get:
      tags: [backups]
      x-role: viewer
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                required: [backups]
                properties:
                  backups: { type: array, items: { $ref: '#/components/schemas/Backup' } }
    post:
      tags: [backups]
      summary: Create a manual backup now
      x-role: operator
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string, maxLength: 200 }
      responses:
        '202': { $ref: '#/components/responses/JobResponse' }
  /instances/{instanceId}/backups/upload:
    post:
      tags: [backups]
      summary: Upload a backup zip previously downloaded from this UI
      x-role: operator
      parameters: [{ $ref: '#/components/parameters/instanceId' }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                required: [backup]
                properties:
                  backup: { $ref: '#/components/schemas/Backup' }
  /instances/{instanceId}/backups/{backupId}:
    parameters:
      - { $ref: '#/components/parameters/instanceId' }
      - { $ref: '#/components/parameters/backupId' }
    delete:
      tags: [backups]
      x-role: operator
      responses:
        '204': { description: Deleted }
  /instances/{instanceId}/backups/{backupId}/download:
    get:
      tags: [backups]
      x-role: operator
      parameters:
        - { $ref: '#/components/parameters/instanceId' }
        - { $ref: '#/components/parameters/backupId' }
      responses:
        '200':
          content:
            application/zip: { schema: { type: string, format: binary } }
  /instances/{instanceId}/backups/{backupId}/restore:
    post:
      tags: [backups]
      summary: Restore a backup (creates a pre_restore backup first; instance must be stopped)
      x-role: operator
      parameters:
        - { $ref: '#/components/parameters/instanceId' }
        - { $ref: '#/components/parameters/backupId' }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StopIfRunning' }
      responses:
        '202': { $ref: '#/components/responses/JobResponse' }
        '409': { $ref: '#/components/responses/Error' }

  # ------------------------------------------------------------------ schedules
...
  /jobs:
    get:
      tags: [jobs]
      x-role: viewer
      parameters:
        - { in: query, name: instance, schema: { type: string } }
        - { in: query, name: status, schema: { $ref: '#/components/schemas/JobStatus' } }
        - { in: query, name: limit, schema: { type: integer, default: 50, maximum: 200 } }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                required: [jobs]
                properties:
                  jobs: { type: array, items: { $ref: '#/components/schemas/Job' } }
  /jobs/{jobId}:
    parameters: [{ $ref: '#/components/parameters/jobId' }]
    get:
      tags: [jobs]
      x-role: viewer
      responses:
        '200': { $ref: '#/components/responses/JobResponse' }
  /jobs/{jobId}/log:
    get:
      tags: [jobs]
      x-role: viewer
      parameters: [{ $ref: '#/components/parameters/jobId' }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                required: [lines]
                properties:
                  lines: { type: array, items: { type: string } }
  /jobs/{jobId}/cancel:
    post:
      tags: [jobs]
      x-role: operator
      parameters: [{ $ref: '#/components/parameters/jobId' }]
      responses:
        '200': { $ref: '#/components/responses/JobResponse' }
        '409': { $ref: '#/components/responses/Error' }

  # ------------------------------------------------------------------ events / audit
...
components:
  parameters:
    instanceId: { in: path, name: instanceId, required: true, schema: { type: string, pattern: '^[a-z0-9][a-z0-9-]{0,31}$' } }
...
    backupId: { in: path, name: backupId, required: true, schema: { type: integer } }
...
    jobId: { in: path, name: jobId, required: true, schema: { type: string, format: uuid } }
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    UserResponse:
      description: User
      content:
        application/json:
          schema:
            type: object
            required: [user]
            properties:
              user: { $ref: '#/components/schemas/User' }
...
    JobResponse:
      description: Job
      content:
        application/json:
          schema:
            type: object
            required: [job]
            properties:
              job: { $ref: '#/components/schemas/Job' }
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: |
                One of: unauthorized, forbidden, not_found, validation_failed, conflict,
                instance_running, instance_not_installed, instance_busy, port_in_use,
                last_admin, invalid_credentials, account_locked, account_disabled,
                setup_done, oidc_disabled, oidc_error, steamcmd_missing, bepinex_missing,
                package_not_found, upstream_error, internal
            message: { type: string }
            details:
              type: object
              additionalProperties: true
              properties:
                fields:
                  type: array
                  items:
                    type: object
                    required: [field, message]
                    properties:
                      field: { type: string }
                      message: { type: string }
    Role:
      type: string
      enum: [viewer, operator, admin]
...
    SetupRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 2, maxLength: 32, pattern: '^[a-z0-9._-]+$' }
        password: { type: string, minLength: 10, maxLength: 128 }
        display_name: { type: string, maxLength: 64 }
        email: { type: string, format: email }
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password: { type: string }
        new_password: { type: string, minLength: 10, maxLength: 128 }
    User:
      type: object
      required: [id, username, role, disabled, has_password, identities, created_at]
      properties:
        id: { type: integer }
        username: { type: string }
        display_name: { type: string }
        email: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        disabled: { type: boolean }
        has_password: { type: boolean, description: false for OIDC-only accounts }
        identities:
          type: array
          items:
            type: object
            required: [provider, subject]
            properties:
              provider: { type: string }
              subject: { type: string }
        created_at: { type: string, format: date-time }
        last_login_at: { type: string, format: date-time }
...
    StopIfRunning:
      type: object
      properties:
        stop_if_running: { type: boolean, default: false, description: Stop the instance for the operation and start it again afterwards }
...
    BackupKind:
      type: string
      enum: [manual, scheduled, pre_update, pre_restore, uploaded]
    Backup:
      type: object
      required: [id, instance_id, world, kind, filename, size_bytes, created_at]
      properties:
        id: { type: integer }
        instance_id: { type: string }
        world: { type: string }
        kind: { $ref: '#/components/schemas/BackupKind' }
        filename: { type: string }
        size_bytes: { type: integer, format: int64 }
        note: { type: string }
        created_at: { type: string, format: date-time }
        missing: { type: boolean, description: DB row exists but file is gone }
...
    JobType:
      type: string
      enum: [install, update, backup, restore, world_import, world_regenerate, mod_install, mod_update, mod_uninstall, bepinex_install, scheduled_restart, thunderstore_refresh, self_upgrade]
    JobStatus:
      type: string
      enum: [queued, running, succeeded, failed, cancelled]
    Job:
      type: object
      required: [id, type, status, created_at]
      properties:
        id: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/JobType' }
        instance_id: { type: string }
        status: { $ref: '#/components/schemas/JobStatus' }
        title: { type: string, description: Human readable, e.g. "Install Jotunn 2.29.2" }
        requested_by: { type: string }
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
        error: { type: string }
        summary:
          type: object
          additionalProperties: true
...